+++ apps/backend/src/modules/contributors/contributor.repository.ts
@@ -1,6 +1,11 @@
 import { raw } from "objection";
 
-import { type Repository } from "~/libs/types/types.js";
+import { SortType } from "~/libs/enums/enums.js";
+import {
+	type PaginationQueryParameters,
+	type PaginationResponseDto,
+	type Repository,
+} from "~/libs/types/types.js";
 import { type GitEmailModel } from "~/modules/git-emails/git-emails.js";
 
 import { ContributorEntity } from "./contributor.entity.js";
@@ -50,11 +55,17 @@ class ContributorRepository implements Repository {
 		return ContributorEntity.initialize(contributor);
 	}
 
-	public async findAll({ hasHidden }: { hasHidden?: boolean } = {}): Promise<{
-		items: ContributorEntity[];
-	}> {
+	public async findAll({
+		hasHidden = false,
+		page,
+		pageSize,
+	}: { hasHidden?: boolean } & PaginationQueryParameters): Promise<
+		PaginationResponseDto<ContributorEntity>
+	> {
 		const query = this.contributorModel
 			.query()
+			.orderBy("createdAt", SortType.DESCENDING)
+			.page(page, pageSize)
 			.select("contributors.*")
 			.select(
 				raw(
@@ -71,12 +82,13 @@ class ContributorRepository implements Repository {
 			query.where("contributors.isHidden", false);
 		}
 
-		const contributorsWithProjectsAndEmails = await query;
+		const { results, total } = await query;
 
 		return {
-			items: contributorsWithProjectsAndEmails.map((contributor) => {
+			items: results.map((contributor) => {
 				return ContributorEntity.initialize(contributor);
 			}),
+			totalItems: total,
 		};
 	}
 
@@ -106,6 +118,40 @@ class ContributorRepository implements Repository {
 
 		return {
 			items: contributorsWithProjectsAndEmails.map((contributor) => {
+				return ContributorEntity.initialize(contributor);
+			}),
+		};
+	}
+
+	public async findAllWithoutPagination({
+		hasHidden = true,
+	}: {
+		hasHidden?: boolean;
+	}): Promise<{
+		items: ContributorEntity[];
+	}> {
+		const query = this.contributorModel
+			.query()
+			.select("contributors.*")
+			.select(
+				raw(
+					"COALESCE(ARRAY_AGG(DISTINCT jsonb_build_object('id', projects.id, 'name', projects.name)) FILTER (WHERE projects.id IS NOT NULL), '{}') AS projects",
+				),
+			)
+			.leftJoin("git_emails", "contributors.id", "git_emails.contributor_id")
+			.leftJoin("activity_logs", "git_emails.id", "activity_logs.git_email_id")
+			.leftJoin("projects", "activity_logs.project_id", "projects.id")
+			.groupBy("contributors.id")
+			.withGraphFetched("gitEmails");
+
+		if (!hasHidden) {
+			query.where("contributors.isHidden", false);
+		}
+
+		const results = await query;
+
+		return {
+			items: results.map((contributor) => {
 				return ContributorEntity.initialize(contributor);
 			}),
 		};
@@ -188,6 +234,46 @@ class ContributorRepository implements Repository {
 			});
 
 		return ContributorEntity.initialize(contributor);
+	}
+
+	public async split(
+		gitEmailId: number,
+		newContributorName: string,
+	): Promise<ContributorEntity> {
+		const result = await this.contributorModel.transaction(async (trx) => {
+			const newContributor = await this.contributorModel
+				.query(trx)
+				.insert({ name: newContributorName })
+				.execute();
+
+			await this.gitEmailModel
+				.query(trx)
+				.patchAndFetchById(gitEmailId, { contributorId: newContributor.id });
+
+			return await this.contributorModel
+				.query(trx)
+				.select("contributors.*")
+				.select(
+					raw(
+						"COALESCE(ARRAY_AGG(DISTINCT jsonb_build_object('id', projects.id, 'name', projects.name)) FILTER (WHERE projects.id IS NOT NULL), '{}') AS projects",
+					),
+				)
+				.leftJoin("git_emails", "contributors.id", "git_emails.contributor_id")
+				.leftJoin(
+					"activity_logs",
+					"git_emails.id",
+					"activity_logs.git_email_id",
+				)
+				.leftJoin("projects", "activity_logs.project_id", "projects.id")
+				.groupBy("contributors.id")
+				.withGraphFetched("gitEmails")
+				.modifyGraph("gitEmails", (builder) => {
+					builder.select("id", "email");
+				})
+				.findById(newContributor.id);
+		});
+
+		return ContributorEntity.initialize(result as ContributorModel);
 	}
 
 	public update(): ReturnType<Repository["update"]> {