+++ apps/backend/src/modules/contributors/contributor.service.ts
@@ -1,6 +1,13 @@
+import {
+	MIN_GIT_EMAILS_LENGTH_FOR_SPLIT,
+	PAGE_INDEX_OFFSET,
+} from "~/libs/constants/constants.js";
 import { ExceptionMessage } from "~/libs/enums/enums.js";
 import { HTTPCode } from "~/libs/modules/http/http.js";
-import { type Service } from "~/libs/types/types.js";
+import {
+	type PaginationQueryParameters,
+	type Service,
+} from "~/libs/types/types.js";
 
 import { ContributorEntity } from "./contributor.entity.js";
 import { type ContributorRepository } from "./contributor.repository.js";
@@ -12,6 +19,7 @@ import {
 	type ContributorMergeRequestDto,
 	type ContributorPatchRequestDto,
 	type ContributorPatchResponseDto,
+	type ContributorSplitRequestDto,
 } from "./libs/types/types.js";
 
 class ContributorService implements Service {
@@ -54,11 +62,15 @@ class ContributorService implements Service {
 
 	public async findAll({
 		hasHidden = true,
+		page,
+		pageSize,
 	}: {
 		hasHidden?: boolean;
-	}): Promise<ContributorGetAllResponseDto> {
+	} & PaginationQueryParameters): Promise<ContributorGetAllResponseDto> {
 		const contributors = await this.contributorRepository.findAll({
 			hasHidden,
+			page: page - PAGE_INDEX_OFFSET,
+			pageSize,
 		});
 
 		return {
@@ -73,6 +85,7 @@ class ContributorService implements Service {
 					})),
 				};
 			}),
+			totalItems: contributors.totalItems,
 		};
 	}
 
@@ -94,6 +107,31 @@ class ContributorService implements Service {
 					})),
 				};
 			}),
+			totalItems: contributors.items.length,
+		};
+	}
+
+	public async findAllWithoutPagination({
+		hasHidden = true,
+	}: {
+		hasHidden?: boolean;
+	}): Promise<ContributorGetAllResponseDto> {
+		const contributors =
+			await this.contributorRepository.findAllWithoutPagination({ hasHidden });
+
+		return {
+			items: contributors.items.map((item) => {
+				const contributor = item.toObject();
+
+				return {
+					...contributor,
+					gitEmails: contributor.gitEmails.map((gitEmail) => ({
+						email: gitEmail.email,
+						id: gitEmail.id,
+					})),
+				};
+			}),
+			totalItems: contributors.items.length,
 		};
 	}
 
@@ -168,6 +206,59 @@ class ContributorService implements Service {
 		}
 
 		return item.toObject();
+	}
+
+	public async split(
+		contributorId: number,
+		payload: ContributorSplitRequestDto,
+	): Promise<ContributorGetAllItemResponseDto | null> {
+		const currentContributorEntity =
+			await this.contributorRepository.find(contributorId);
+		const hasContributor = currentContributorEntity !== null;
+
+		if (!hasContributor) {
+			throw new ContributorError({
+				message: ExceptionMessage.CONTRIBUTOR_NOT_FOUND,
+				status: HTTPCode.NOT_FOUND,
+			});
+		}
+
+		const currentContributor = currentContributorEntity.toObject();
+
+		if (
+			currentContributor.gitEmails.length <= MIN_GIT_EMAILS_LENGTH_FOR_SPLIT
+		) {
+			throw new ContributorError({
+				message: ExceptionMessage.CONTRIBUTOR_SPLIT_SINGLE_EMAIL,
+				status: HTTPCode.CONFLICT,
+			});
+		}
+
+		const splitEmail = currentContributor.gitEmails.find(
+			({ id }) => id === payload.gitEmailId,
+		);
+		const hasSplittedEmail = splitEmail !== undefined;
+
+		if (!hasSplittedEmail) {
+			throw new ContributorError({
+				message: ExceptionMessage.CONTRIBUTOR_SPLIT_FAILED,
+				status: HTTPCode.CONFLICT,
+			});
+		}
+
+		try {
+			const splitContributor = await this.contributorRepository.split(
+				payload.gitEmailId,
+				payload.newContributorName,
+			);
+
+			return splitContributor.toObject();
+		} catch {
+			throw new ContributorError({
+				message: ExceptionMessage.CONTRIBUTOR_SPLIT_FAILED,
+				status: HTTPCode.CONFLICT,
+			});
+		}
 	}
 
 	public update(): ReturnType<Service["update"]> {