+++ apps/frontend/src/libs/types/icon-name.type.ts
@@ -17,6 +17,7 @@ type IconName =
 	| "rightArrow"
 	| "rightDoubleArrow"
 	| "search"
+	| "split"
 	| "strikedEye"
 	| "trashBin"
 	| "warning";